+++ Vanilla option pricer/matrix_plf.cpp
@@ -4,15 +4,15 @@
 #include <memory>
 
 
-double matrix_plf::operator()(double t) const {
-    int columns = get_number_of_columns();
-    // The operator returns the first (last) value of returns if the double t is lower than the lowest time (is higher than the highest time) in the tab
+  long double  matrix_plf::operator()(  long double  t) const {
+    unsigned int columns = get_number_of_columns();
+    // The operator returns the first (last) value of returns if the   long double  t is lower than the lowest time (is higher than the highest time) in the tab
     if (t <= matrix::operator()(1, 1)) { return matrix::operator()(2, 1); }
     else if (t >= matrix::operator()(1, columns)) { return matrix::operator()(2, columns); }
     else {
         // Linear interpolation for t in [t_i; t_i+1]
-        double t1, t2, r1, r2;
-        for (int j = 1; j < columns; ++j) {
+          long double  t1, t2, r1, r2;
+        for (unsigned int j = 1; j < columns; ++j) {
             t1 = matrix::operator()(1, j);
             t2 = matrix::operator()(1, j + 1);
             if (t >= t1 && t <= t2) {
@@ -27,15 +27,15 @@ double matrix_plf::operator()(double t) const {
 }
 
 
-matrix_plf matrix_plf::shift_value(double delta_r){
-    int numberOfPoints = get_number_of_columns();
+matrix_plf matrix_plf::shift_value(  long double  delta_r){
+    unsigned int numberOfPoints = get_number_of_columns();
 
     // Create 2 smart pointers for arrays of times and returns
-    auto tab_t = std::make_unique<double[]>(numberOfPoints);
-    auto tab_r = std::make_unique<double[]>(numberOfPoints);
+    auto tab_t = std::make_unique< long double []>(numberOfPoints);
+    auto tab_r = std::make_unique< long double []>(numberOfPoints);
 
     // Fill `tab_t` and `tab_r` with the current values
-    for (int j = 1; j <= numberOfPoints; j++) {
+    for (unsigned int j = 1; j <= numberOfPoints; j++) {
         tab_t[j - 1] = matrix::operator()(1, j);          // Copy the first row (t values)
         tab_r[j - 1] = matrix::operator()(2, j) + delta_r; // Shift the second row (r values) by `delta_r`
     }
@@ -47,12 +47,12 @@ matrix_plf matrix_plf::shift_value(double delta_r){
 }
 
 
-double matrix_plf::integral(double t, double T) const
+  long double  matrix_plf::integral(  long double  t,   long double  T) const
 {
-    int i = 1;
+    unsigned int i = 1;
     while (i + 1 < get_number_of_columns() && matrix::operator()(1, i + 1) < t) i++;
-    double res = 0.0;
-    double r = (*this)(t);
+      long double  res = 0.0;
+      long double  r = (*this)(t);
     res += (matrix::operator()(2, i + 1) + r) * (matrix::operator()(1, i + 1) - t) / 2;
     i++;
