+++ Vanilla option pricer/matrix_plf.h
@@ -33,7 +33,7 @@ public:
 	 * @param numberOfPoints The number of points in the matrix.
 	 * Initializes a 2-row matrix with the given number of columns.
 	 */
-	matrix_plf(int numberOfPoints) : matrix(2, numberOfPoints) {}
+	matrix_plf(unsigned int numberOfPoints) : matrix(2, numberOfPoints) {}
 
 	/**
 	 * @brief Constructs a `matrix_plf` object from input arrays with time-value pairs.
@@ -47,9 +47,9 @@ public:
 	 * - Row 1 contains elements from `tab_t`.
 	 * - Row 2 contains elements from `tab_r`.
 	 */
-	matrix_plf(const double* tab_t, const double* tab_r, int numberOfPoints) : matrix(2, numberOfPoints) {
+	matrix_plf(const  long double* tab_t, const  long double* tab_r, unsigned int numberOfPoints) : matrix(2, numberOfPoints) {
 		// Directly access elements with 1-based indexing
-		for (int j = 1; j <= numberOfPoints; ++j) {
+		for (unsigned int j = 1; j <= numberOfPoints; ++j) {
 			matrix::operator()(1, j) = tab_t[j - 1]; // Fill row 1 with `tab_t`
 			matrix::operator()(2, j) = tab_r[j - 1]; // Fill row 2 with `tab_r`
 		}
@@ -64,7 +64,7 @@ public:
 	 * - The value at the first time point if `t` is before the first date.
 	 * - The value at the last time point if `t` is after the last date.
 	 */
-	double operator()(double t) const;
+	 long double operator()( long double t) const;
 
 	/**
 	 * @brief Shifts all values in the matrix by a specified delta.
@@ -72,7 +72,7 @@ public:
 	 * @param delta The value by which to shift the risk-free rate.
 	 * @return A new `matrix_plf` object with all values shifted by `delta`.
 	 */
-	matrix_plf shift_value(double delta);
+	matrix_plf shift_value( long double delta);
 	
 	/**
 	 * @brief Computes the discount rate between two time points.
@@ -81,5 +81,5 @@ public:
 	 * @param T The end time.
 	 * @return The computed discount rate between `t` and `T`.
 	 */
-	double integral(double t, double T) const;
+	 long double integral( long double t,  long double T) const;
 };
