+++ Vanilla option pricer/matrix_tridiagonal.cpp
@@ -2,40 +2,40 @@
 #include "matrix_tridiagonal.h"
 
 matrix matrix_tridiagonal::inverse() const {
-    int dim = get_number_of_lines();
+    unsigned int dim = get_number_of_lines();
 
     // Copy of the current matrix to work on for inversion
     matrix m = matrix(dim, dim);
-    for (int i = 1; i <= dim; ++i) { for (int j = 1; j <= dim; ++j) { m(i, j) = (*this)(i, j); } }
+    for (unsigned int i = 1; i <= dim; ++i) { for (unsigned int j = 1; j <= dim; ++j) { m(i, j) = (*this)(i, j); } }
 
     // Initialize the identity matrix to store the inverse
     matrix m_1(dim, dim);
-    for (int i = 1; i <= dim; ++i) {
+    for (unsigned int i = 1; i <= dim; ++i) {
         m_1(i, i) = 1.0;
     }
 
     // Lower triangular elimination
-    for (int i = 1; i < dim; ++i) {
-        double mii = m(i, i);
-        double alpha = m(i + 1, i);
+    for (unsigned int i = 1; i < dim; ++i) {
+          long double  mii = m(i, i);
+          long double  alpha = m(i + 1, i);
 
         m.add_line_to_line(i + 1, -alpha / mii, i);
         m_1.add_line_to_line(i + 1, -alpha / mii, i);
     }
 
     // Upper triangular elimination
-    for (int i = dim; i > 1; --i) {
-        double mii = m(i, i);
-        double alpha = m(i - 1, i);
+    for (unsigned int i = dim; i > 1; --i) {
+          long double  mii = m(i, i);
+          long double  alpha = m(i - 1, i);
 
         m.add_line_to_line(i - 1, -alpha / mii, i);
         m_1.add_line_to_line(i - 1, -alpha / mii, i);
     }
 
     // Normalize diagonal to 1
-    for (int i = 1; i <= dim; ++i) {
-        double mii = m(i, i);
-        for (int j = 1; j <= dim; ++j) {
+    for (unsigned int i = 1; i <= dim; ++i) {
+          long double  mii = m(i, i);
+        for (unsigned int j = 1; j <= dim; ++j) {
             m_1(i, j) /= mii;
         }
     }
@@ -44,35 +44,35 @@ matrix matrix_tridiagonal::inverse() const {
 
 matrix matrix_tridiagonal::gauss_saidel(matrix b, matrix x0) const {
 
-    int dim = b.get_number_of_lines();
+    unsigned int dim = b.get_number_of_lines();
     matrix x = x0;      // Initialize x0
 
     // Repetition for convergence
-    const int max_iter = 1000; // Max number of iteration 
-    const double tol = 1e-12;   // Tollerance for convergence 
-    for (int iter = 0; iter < max_iter; ++iter) {
-        double max_diff = 0.0; // Difference between old and new x
+    const unsigned int max_iter = 1000; // Max number of iteration 
+    const   long double  tol = 1e-12;   // Tollerance for convergence 
+    for (unsigned int iter = 0; iter < max_iter; ++iter) {
+          long double  max_diff = 0.0; // Difference between old and new x
 
-        for (int i = 1; i <= dim; ++i) {
-            double sum = b(i, 1); // Initialize the term
+        for (unsigned int i = 1; i <= dim; ++i) {
+              long double  sum = b(i, 1); // Initialize the term
 
             // Diagonal element
-            double a_ii = (*this)(i, i);
+              long double  a_ii = (*this)(i, i);
 
             // Sum of elements on the left
             if (i > 1) {
-                double a_ij = (*this)(i, i - 1);
+                  long double  a_ij = (*this)(i, i - 1);
                 sum -= a_ij * x(i - 1, 1);
             }
 
             // Sum of the elements on the right
             if (i < dim) {
-                double a_ij = (*this)(i, i + 1);
+                  long double  a_ij = (*this)(i, i + 1);
                 sum -= a_ij * x0(i + 1, 1);
             }
 
             // Update the solution
-            double x_new = sum / a_ii;
+              long double  x_new = sum / a_ii;
             max_diff = (max_diff > std::abs(x_new - x(i, 1))) ? max_diff : std::abs(x_new - x(i, 1));
             x(i, 1) = x_new;
         }
