+++ Vanilla option pricer/pricer.cpp
@@ -4,27 +4,27 @@
 #include "matrix_plf.h"
 #include "matrix_tridiagonal.h"
 #include "mesh_matrix.h"
-double computePriceAndPartialDerivates(
-    double* result,
-    double* underlying,
-    double* option_price,
-    double* delta,
-    double* exerciceBoundary,
+  long double  computePriceAndPartialDerivates(
+      long double * result,
+      long double * underlying,
+      long double * option_price,
+      long double * delta,
+      long double * exerciceBoundary,
     bool isCall,
     bool isEuropean,
-    double maturity,
-    double strike,
-    int time_mesh_params,
-    int spot_mesh_params,
-    double S0,
-    const double* time_steps,  // Time steps array
-    const double* r_values,   // Corresponding interest rates array
-    int size,                 // Size of the r table (number of time steps)
-    double sigma)
+      long double  maturity,
+      long double  strike,
+    unsigned int time_mesh_params,
+    unsigned int spot_mesh_params,
+      long double  S0,
+    const   long double * time_steps,  // Time steps array
+    const   long double * r_values,   // Corresponding interest rates array
+    unsigned int size,                 // Size of the r table (number of time steps)
+      long double  sigma)
 {
 
-    double sinf = S0 * 2.5;
-    long double ds = sinf / (spot_mesh_params - 1);
+      long double  sinf = S0 * 2.5;
+       long double  ds = sinf / (spot_mesh_params - 1);
 
     matrix_plf returnTable = matrix_plf(time_steps, r_values, size);
     mesh_matrix m_matrix(maturity, S0, strike, spot_mesh_params, time_mesh_params, sinf);
@@ -35,19 +35,19 @@ double computePriceAndPartialDerivates(
     underlying[0] = 0.0;
     option_price[0] = m_matrix(1, 1);
 
-    for (int j = 2; j < spot_mesh_params - 2; ++j) {
+    for (unsigned int j = 2; j < spot_mesh_params - 2; ++j) {
         underlying[j - 1] = (j - 1) * ds;
         delta[j - 1] = m_matrix.retrieve_delta((j - 1) * ds);
         option_price[j - 1] = m_matrix(1, j);
     }
 
-    for (int i = 1; i <= time_mesh_params - 2; ++i) {
+    for (unsigned int i = 1; i <= time_mesh_params - 2; ++i) {
         exerciceBoundary[i - 1] = isCall ? sinf : 0.0;
 
-        for (int jj = isCall ? 1 : spot_mesh_params;
+        for (unsigned int jj = isCall ? 1 : spot_mesh_params;
             isCall ? (jj <= spot_mesh_params) : (jj >= 1);
             jj += isCall ? 1 : -1) {
-            double S = sinf * (jj - 1) / (spot_mesh_params - 1);
+              long double  S = sinf * (jj - 1) / (spot_mesh_params - 1);
             if ((isCall ? (S - strike) : (strike - S)) - m_matrix(i, jj) == 0) {
                 exerciceBoundary[i - 1] = S;
                 break;
